chore: move market link into vega entities file
