+++ frontend/components/receipts/utils/vega-entities/market-link.tsx
